Reorganize validator types and remove backup file
